// src/app/model/ISelectSiniestro.ts
export class ISelectSiniestro {
  idsiniestro: string = "";
  Persona: string = "";
  vplaca: string = "";
  dFecRegistro: string = "";
  Estado: string = "";
  TotalRegistros: string = "";
  smidciaseguros: string = "";
  idpoliza: string = "";
  planPoliza: string = "";
  idvehiculo: string = "";
  iestadosiniestro: string = "";
  dFecNotificacion: string = "";
  dUltmodificación: string = "";
  idconsecuencia: string = "";
  dFecOcurrencia: string = "";
  vlugarsiniestro: string = "";
  vubicasiniestro: string = "";
  iocupantes: string = "";
  idtipodeclarante: string = "";
  vdenominacion: string = "";
  vtelef_declarante: string = "";
  iparentaseg_declarante: string = "";
  vmaildeclarante: string = "";
  vconductor: string = "";
  idtipodoc: string = "";
  vnrodociden: string = "";
  vlicencia: string = "";
  iparentaseg_conductor: string = "";
  vtelef_conductor: string = "";
  vemail_conductor: string = "";
  dvencilicencia: string = "";
  idcomisaria: string = "";
  descomisaria: string = "";
  vcategoria: string = "";
  vdetasiniestro: string = "";
  nidusuario: string = "";
  vimgtarjetaprop: string = "";
  vimglicconducir: string = "";
  vimgsoat: string = "";
  idgravedad: string = "";
  smcausa: string = "";
  idetallecausa: string = "";
  idtipo_decla: string = "";
  idtipo_docid_decla: string = "";
  idnrodoc_decla: string = "";
  vnombres_decla: string = "";
  vapellidopat_decla: string = "";
  vapellidomat_decla: string = "";
  idtipo_conduc: string = "";
  idtipo_docid_conduc: string = "";
  idnrodoc_conduc: string = "";
  vnombres_conduc: string = "";
  vapellidopat_conduc: string = "";
  vapellidomat_conduc: string = "";
  smidubigeo: string = "";
  vdescripcionubi: string = "";
  dFecDenunciaP: string = "";
  dFecCambioEstado: string = "";
  dUser: string = "";
  placa: string = "";
  estado: string = "";
  vigencia: string = "";
  Clase: string = "";
  Marca: string = "";
  Modelo: string = "";
  Vin: string = "";
  Anio: string = "";
  Asientos: string = "";
  Kilometraje: string = "";
  idanalista: string = "";
  analista: string = "";
  idsiniestro_ou: string = "";
  smcausades: string = "";
  iddetalledes: string = "";
  idaseguradora: string = "";
  dfechaReporteProc: string = "";
  dfechaInicioProc: string = "";
  dfechaFinProc: string = "";
  dtipoasesoria: string = "";
  dkilometraje: string = "";
  idsemaforo: string = "";
  idplan: string = "";
  idnrodocumento: string = "";
  vcelular: string = "";
  vemail: string = "";
  horaocurrencia: string = "";
  marca: string = "";
  modelo: string = "";
  color: string = "";
  nrovin: string = "";
  kilometraje: string = "";
  EstadoSini: string = "";
  descripcion: string = "";
  observacion: string = "";
  daniosauto: string = "";
  dfechaAsignacion: string = "";
  idprocurador: string = "";
  procurador: string = "";
  estadopoliza: string = "";
  icodigorenovacion: string = "";

  dFecModificacion: string = "";
  Uso: string = "";

  smcausageneral: string = "";
  descripcionimagen: string = "";
  descripcionseccion: string = "";
  Moneda: string = "";
  sumaAsegurada: string = "";
  dcminimo: string = "";
  porcentajeDedu: string = "";
  reserinireg: string = "";
  perdidaTotal: boolean = false;

  personaPC: string = "";
  idpersonaTP: string = "";
  flagAsegurado: boolean = false;
  flagConductor : boolean = false;

}